Client: Automated BDB Environment placing
Instead of requiring a script (or human intervention) to place BDB's Environment, it is instead initalized to a folder in the current user's home directory called ".instadisc". This way, not only can each user on a computer can use InstaDisc seperatly, this rids the need to provide command-line arguments to InstaDisc (with the exception of -r, for external invoking of retained requesting).