Client: Fixed up First Run form
Changed the password field (which was a JTextField) in Step 2 to a JPasswordField which has getText() deprecated in favor of getPassword(), which returns the password in a char[]. Since I currently cannot be bothered parsing a char[], I've left it using the deprectated getText(). This must be fixed at some point.