Central: Fixed double-sending issue
Previously, it was thought that the Client was spawing two identical
HandleItemThreads, but the real reason was that the Central Server was sending
the item twice because the subscribement was in the database twice, once
because for ownership, the second for subscribement. Now the Central Server
has been fixed to prevent this error when a user is subscribed to a
subscription they own. Fixes #15.