Update: Created Subscription Files
Currently, a seperate php file generates the subscription files. This is a bad scheme as it requires
the subscription URL (as in, the URL you use to subscribe, the URL to the Subscription File) is long
and winds down the Wordpress plugin directory. Hopefully, an alternate to this method will be found.
Refs #6.