Client: Fixed yet another client item bug
This once was caused by a combination of having more than one subscription in the database and a bug in the subscription retrieval method. The bug has been fixed which allows us to add back the Item buffer (removed in [68] because it was thought to be the cause of another Client Item bug) Fixes #18.