Update: Worked on phpBB3 plugin
The Subscription File didn't contain the activation key, so that was fixed, and
also it turned out that everywhere was using the wrong method of getting the
board's URL, so that was changed too. Then then sending library was worked
upon. The only fault I currently see is that the Author field doesn't seem to
be being sent for some reason. Refs #7