Central: Escape ": "s
If a colon is before a space in any value sent from the Central Server to the
Client, it will be interpreted as a NAME-VALUE divider and mess up the Item.
So, any wild colon-spaces have been escaped as "__INSTADISC__". Next step is to
unescape them when they reach the Client.