Client: Transactionalized the database
Now, the database is ACID-compliant. Also fixed the error that occurs when you get sent an item you don't have a subscription for, and it tries to delete the subscription's entry off of the Central Server and fails. The error was because the Client was attempting to look up the Subscription's URL in the database after it already had the URL and ensured that the Subscription wasn't there.