Update: Seperated Encryption ID from Verifiction ID
The Client was failing to properly decrypt encrypted data because of a key
generation issue, the key is generated from a password and a Verification ID.
However, because Verifcation IDs must be unique, the encryption key and the
decryption key would differ. However, the Central Server needs to be modified
so as to pass on the Encryption Key. Refs #10