Central: Moved over to Four Island Templating system
Borrowed the Four Island templating system code so as to ease the creation of frontends. index.php and activate.php have already been converted, and install.php will not
be converted (because it would be too difficult, and who would want to skin it anyway?)

NOTICE: The Four Island Templating system source code is copyrighted by Starla Insigna (c) 2008. All rights reserved. Use of this code in part or in whole apart from the
InstaDisc Central Server is a direct violation of copyright law. You have been warned.

Refs #25