General: Upped default verIDBufferSize
Because of the increase in possible Verification ID values, it will be even less likely to have a repeat, so the number that are buffered is being increased. A test was
run to find repeats if 10000 numbers were generated between 1 and 2147483647 and a repeat was found only once at the 8500th number, the rest of the tests (20 others)
passed successfully. Refs #37